Set mediarequests per referer requests field as long
It was accidentally changed to int in the last patch.

Bug: T233622

+++ sys/mediarequests.js
@@ -62,7 +62,7 @@ const tableSchemas = {
             agent: 'string',
             granularity: 'string',
             timestamp: 'string',
-            requests: 'int'
+            requests: 'long'
         },
         index: [
             { attribute: 'referer', type: 'hash' },
